refactor(421): extract stateless host families (random, math) to vera/runtime/
Step 4 of the codegen/api.py decomposition (#421): the two stateless host-binding families -- Random (#465) and Math (#467) -- move out of execute() into vera/runtime/random.py and vera/runtime/math.py, each exposing a register_<family>(linker, ops_used) that defines and registers its host callbacks. execute() now calls register_random/register_math instead of inlining the blocks. This templates the per-family pattern for the stateful families to follow. api.py 3,001->2,884 lines. mypy + ruff clean; the random/math codegen tests (27) + #734 gate stay green.
